+++ src/main/java/net/firestaff/mcp/baselab/shapes/complex/TranslatedQuarterHalfedComplexShape.java
@@ -57,8 +57,8 @@ public class TranslatedQuarterHalfedComplexShape extends AbstractComplexShape im
         }
 
 
-        p.x = (int)x;
-        p.y = (int)y;
+        p.x = x;
+        p.y = y;
         p = Rotate.rotate(p, Math.PI/2);
 
         return p;
