+++ src/main/java/net/firestaff/mcp/baselab/shapes/secondorder/EpicycloidThree.java
@@ -12,6 +12,6 @@ public class EpicycloidThree extends Epicycloid {
 
     public Point draw(double length, double circleModifier, double circleAngle, double circleTurnModifier) {
         Point p = drawPetals(length,  circleModifier,  circleAngle + Math.PI,  circleTurnModifier,  3);
-        return Rotate.rotate(new Point((int)p.x,(int)p.y), Math.PI);
+        return Rotate.rotate(new Point(p.x,p.y), Math.PI);
     }
 }
