+++ src/main/java/net/firestaff/mcp/baselab/shapes/secondorder/HypocycloidReal.java
@@ -34,8 +34,8 @@ public class HypocycloidReal  extends AbstractShape implements Shape {
         // swapping both makes for turned hypocycloid.
         p = Rotate.rotate(p, Math.PI / 2);
 
-        p.y = (int)y;
-        p.x = (int)x;
+        p.y = y;
+        p.x = x;
 
         return p;
     }
@@ -53,8 +53,8 @@ public class HypocycloidReal  extends AbstractShape implements Shape {
         // swapping both makes for turned hypocycloid.
         p = Rotate.rotate(p, Math.PI / 2);
 
-        p.y = (int)y;
-        p.x = (int)x;
+        p.y = y;
+        p.x = x;
 
         return p;
     }
