+++ src/main/java/net/firestaff/mcp/baselab/shapes/secondorder/Hypotrochoid.java
@@ -32,8 +32,8 @@ public class Hypotrochoid extends AbstractShape implements Shape {
         // swapping both makes for turned hypocycloid.
         p = Rotate.rotate(p, Math.PI / 2);
 
-        p.y = (int)y;
-        p.x = (int)x;
+        p.y = y;
+        p.x = x;
 
         return p;
     }
