+++ src/main/java/net/firestaff/mcp/baselab/shapes/weird/PlusHalf.java
@@ -26,7 +26,7 @@ public class PlusHalf extends AbstractShape implements Shape {
             else {
                 x = 0;
                 y = ((length / 2) * circleModifier) * Math.sin((circleAngle - (Math.PI / 2)));
-                //p = reverse(new Point((int)x,(int)y)); x = p.x; y = p.y;
+                //p = reverse(new Point(x,y)); x = p.x; y = p.y;
             }
         }
         else if(circleAngle < (Math.PI ))
@@ -38,7 +38,7 @@ public class PlusHalf extends AbstractShape implements Shape {
             else {
                 x = ((length / 2) * circleModifier) * Math.sin((circleAngle - (Math.PI / 2)));
                 y = 0;
-                //p = reverse(new Point((int)x,(int)y)); x = p.x; y = p.y;
+                //p = reverse(new Point(x,y)); x = p.x; y = p.y;
             }
         }
         else if(circleAngle < (Math.PI + (Math.PI / 2)))
@@ -74,8 +74,8 @@ public class PlusHalf extends AbstractShape implements Shape {
                 y = ((length / 2) * circleModifier) * Math.sin((circleAngle + Math.PI / 2));
             }
         }
-        p.x = (int)x;
-        p.y = (int)y;
+        p.x = x;
+        p.y = y;
 
         return p;
     }
